+++ Source/Tools/FEXLoader/TestHarnessRunner.cpp
@@ -207,7 +207,7 @@ int main(int argc, char **argv, char **const envp) {
   LogMan::Msg::InstallHandler(MsgHandler);
   FEXCore::Config::Initialize();
   FEXCore::Config::AddLayer(fextl::make_unique<FEX::ArgLoader::ArgLoader>(argc, argv));
-  FEXCore::Config::AddLayer(FEXCore::Config::CreateEnvironmentLayer(envp));
+  FEXCore::Config::AddLayer(FEX::Config::CreateEnvironmentLayer(envp));
   FEXCore::Config::Load();
 
   auto Args = FEX::ArgLoader::Get();
